+++ src/scaffold/ivy/src/app/app.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+
+import { AppComponent } from './app.component';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }