+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.0.0-rc.6
 
+- NEW `--ivy` argument will scaffold with a minimal setup for angular ivy
 - FIXED issue with sourcemaps when using `--env` in dev build
 
 -------------------------------------------------------------------------------------------------------------